+++ libcaf_core/caf/flow/op/cell.hpp
@@ -185,7 +185,7 @@ private:
     }
     if (out_) {
       if (from_external)
-        out_.on_complete();
+        out_.on_error(make_error(sec::disposed));
       else
         out_.release_later();
     }