+++ libcaf_core/caf/flow/op/concat.hpp
@@ -165,7 +165,7 @@ private:
     sub_.cancel();
     fwd_sub_.cancel();
     if (from_external)
-      out_.on_complete();
+      out_.on_error(make_error(sec::disposed));
     else
       out_.release_later();
   }