+++ libcaf_core/caf/flow/op/interval.cpp
@@ -84,7 +84,7 @@ private:
     }
     pending_.dispose();
     if (from_external)
-      out_.on_complete();
+      out_.on_error(make_error(sec::disposed));
     else
       out_.release_later();
   }