+++ libcaf_core/caf/flow/op/merge.hpp
@@ -240,7 +240,7 @@ private:
       kvp.second.sub.cancel();
     sub_.cancel();
     if (from_external)
-      out_.on_complete();
+      out_.on_error(make_error(sec::disposed));
     else
       out_.release_later();
   }