+++ libcaf_core/caf/flow/op/never.hpp
@@ -42,7 +42,7 @@ private:
     if (!out_)
       return;
     if (from_external)
-      out_.on_complete();
+      out_.on_error(make_error(sec::disposed));
     else
       out_.release_later();
   }